+++ hasher-matcher-actioner/hmalib/common/tests/test_content_models.py
@@ -12,7 +12,8 @@ from hmalib.common.models.content import (
     ContentRefType,
 )
 from hmalib.common.configs.evaluator import ActionLabel, ActionRule
-from hmalib.common.message_models import BankedSignal, ActionMessage
+from hmalib.common.messages.action import ActionMessage
+from hmalib.common.messages.match import BankedSignal
 from hmalib.common.classification_models import ClassificationLabel
 from hmalib.common.configs.actioner import WebhookPostActionPerformer
