+++ hasher-matcher-actioner/hmalib/lambdas/actions/action_performer.py
@@ -6,7 +6,7 @@ import datetime
 from functools import lru_cache
 from mypy_boto3_dynamodb import DynamoDBServiceResource
 
-from hmalib.common.message_models import ActionMessage
+from hmalib.common.messages.action import ActionMessage
 from hmalib.common.configs.actioner import ActionPerformer
 from hmalib.common.models.content import ActionEvent
 from hmalib.common.logging import get_logger
